requirements.txt
uvicorn==0.11.8
fastapi-utils==0.2.1
